+++ mesonbuild/backend/ninjabackend.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 from . import backends
+from .. import modules
 from .. import environment, mesonlib
 from .. import build
 from .. import mlog
@@ -1034,7 +1035,7 @@ int dummy;
                 args += d.get_lang_args('vala')
         # Detect gresources and add --gresources arguments for each
         for (gres, gensrc) in other_src[1].items():
-            if hasattr(gensrc, 'gresource_c_output'):
+            if isinstance(gensrc, modules.GResourceTarget):
                 gres_xml, = self.get_custom_target_sources(gensrc)
                 args += ['--gresources=' + gres_xml]
         extra_args = []
